Refactor field names and change packages
